fix(cli/config): fixed trying to link share/man when ln already exists

+++ cli/config.go
@@ -64,8 +64,10 @@ func createPack() error {
 		return err
 	}
 
-	if err := exec.RunInPack("share", []string{"ln -s ../man man"}); err != nil {
-		return err
+	if !util.PathExists(util.PathPackDir("share/man")) {
+		if err := exec.RunInPack("share", []string{"ln -s ../man man"}); err != nil {
+			return err
+		}
 	}
 
 	return nil
